+++ examples/variables_types_values.hlts
