+++ src/AST.hpp
@@ -1,448 +1,14 @@
 #pragma once
 
-#include <string>
-#include <vector>
-#include <iostream>
-#include <memory>
-
-// Data types supported by Helets
-enum class DataType
-{
-    INT,
-    FLOAT,
-    BOOL,
-    STRING
-};
-
-// Operator types for binary operations
-enum class OpType
-{
-    ADD,
-    SUB,
-    MUL,
-    DIV,
-    VAL_EQ,  // =
-    VAL_NEQ, // !=
-    GT,  // >
-    GTE, // >=
-    LT,  // <
-    LTE  // <=
-};
-
-// Forward declarations
-class ExprNode;
-class StmtNode;
-
-// Base node for the AST
-class ASTNode
-{
-  public:
-    virtual ~ASTNode() = default;
-    virtual void print(int indent = 0) const = 0;
-
-  protected:
-    void printIndent(int indent) const
-	{
-        for (int i = 0; i < indent; ++i) {
-            std::cout << "  ";
-        }
-    }
-};
-
-// Expression node (base class for all expressions)
-class ExprNode : public ASTNode
-{
-  public:
-    virtual DataType getType() const = 0;
-};
-
-// Integer literal node
-class IntLiteralNode : public ExprNode
-{
-  public:
-    IntLiteralNode(int value) : value(value) {}
-
-    void print(int indent = 0) const override
-	{
-        printIndent(indent);
-        std::cout << "IntLiteral: " << value << std::endl;
-    }
-
-    DataType getType() const override
-	{
-        return DataType::INT;
-    }
-
-    int getValue() const
-	{
-        return value;
-    }
-
-  private:
-    int value;
-};
-
-// Float literal node
-class FloatLiteralNode : public ExprNode
-{
-  public:
-    FloatLiteralNode(float value) : value(value) {}
-
-    void print(int indent = 0) const override
-	{
-        printIndent(indent);
-        std::cout << "FloatLiteral: " << value << std::endl;
-    }
-
-    DataType getType() const override
-	{
-        return DataType::FLOAT;
-    }
-
-    float getValue() const
-	{
-        return value;
-    }
-
-  private:
-    float value;
-};
-
-// Boolean literal node
-class BoolLiteralNode : public ExprNode
-{
-  public:
-    BoolLiteralNode(bool value) : value(value) {}
-
-    void print(int indent = 0) const override
-	{
-        printIndent(indent);
-        std::cout << "BoolLiteral: " << (value ? "true" : "false") << std::endl;
-    }
-
-    DataType getType() const override
-	{
-        return DataType::BOOL;
-    }
-
-    bool getValue() const
-	{
-        return value;
-    }
-
-private:
-    bool value;
-};
-
-// String literal node
-class StringLiteralNode : public ExprNode
-{
-public:
-    StringLiteralNode(const char* value) : value(value) {}
-    ~StringLiteralNode()
-	{
-        free(const_cast<char*>(value));
-    }
-
-    void print(int indent = 0) const override
-	{
-        printIndent(indent);
-        std::cout << "StringLiteral: " << value << std::endl;
-    }
-
-    DataType getType() const override
-	{
-        return DataType::STRING;
-    }
-
-    const char* getValue() const
-	{
-        return value;
-    }
-
-  private:
-    const char* value;
-};
-
-// Identifier node (for variable references)
-class IdentifierNode : public ExprNode
-{
-public:
-    IdentifierNode(const char* name) : name(name) {}
-    ~IdentifierNode()
-    {
-        free(const_cast<char*>(name));
-    }
-
-    void print(int indent = 0) const override
-    {
-        printIndent(indent);
-        std::cout << "Identifier: " << name << std::endl;
-    }
-
-    DataType getType() const override
-    {
-        // In a real implementation, we would look up the identifier's type
-        // For now, default to INT
-        return DataType::INT;
-    }
-
-    const char* getName() const
-    {
-        return name;
-    }
-
-private:
-    const char* name;
-};
-
-// Binary operation node (for arithmetic operations)
-class BinaryOpNode : public ExprNode
-{
-  public:
-    BinaryOpNode(OpType op, ExprNode* left, ExprNode* right)
-        : op(op), left(left), right(right) {}
-
-    ~BinaryOpNode()
-	{
-        delete left;
-        delete right;
-    }
-
-    void print(int indent = 0) const override
-	{
-        printIndent(indent);
-        std::cout << "BinaryOp: ";
-        switch (op) {
-            case OpType::ADD: std::cout << "+"; break;
-            case OpType::SUB: std::cout << "-"; break;
-            case OpType::MUL: std::cout << "*"; break;
-            case OpType::DIV: std::cout << "/"; break;
-            case OpType::VAL_EQ: std::cout << "="; break;
-            case OpType::VAL_NEQ: std::cout << "!="; break;
-            case OpType::GT: std::cout << ">"; break;
-            case OpType::GTE: std::cout << ">="; break;
-            case OpType::LT: std::cout << "<"; break;
-            case OpType::LTE: std::cout << "<="; break;
-        }
-        std::cout << std::endl;
-
-        printIndent(indent);
-        std::cout << "Left:" << std::endl;
-        left->print(indent + 1);
-
-        printIndent(indent);
-        std::cout << "Right:" << std::endl;
-        right->print(indent + 1);
-    }
-
-    DataType getType() const override
-	{
-        // For comparison operators, the result is always boolean
-        switch (op) {
-            case OpType::VAL_EQ:
-            case OpType::VAL_NEQ:
-            case OpType::GT:
-            case OpType::GTE:
-            case OpType::LT:
-            case OpType::LTE:
-                return DataType::BOOL;
-            default:
-                break;
-        }
-
-        // For arithmetic operators:
-        DataType leftType = left->getType();
-        DataType rightType = right->getType();
-
-        // If either operand is a float, the result is a float
-        if (leftType == DataType::FLOAT || rightType == DataType::FLOAT)
-		{
-            return DataType::FLOAT;
-        }
-
-        // Otherwise, inherit the type from the left operand
-        return leftType;
-    }
-
-    OpType getOp() const
-	{
-        return op;
-    }
-
-    ExprNode* getLeft() const
-	{
-        return left;
-    }
-
-    ExprNode* getRight() const
-	{
-        return right;
-    }
-
-  private:
-    OpType op;
-    ExprNode* left;
-    ExprNode* right;
-};
-
-// Statement node (base class for all statements)
-class StmtNode : public ASTNode{
-};
-
-// Variable declaration node
-class VarDeclNode : public StmtNode
-{
-public:
-    VarDeclNode(DataType type, const char* name, ExprNode* value)
-        : type(type), name(name), value(value) {}
-
-    ~VarDeclNode()
-	{
-        free(const_cast<char*>(name));
-        delete value;
-    }
-
-    void print(int indent = 0) const override
-	{
-        printIndent(indent);
-        std::cout << "VarDecl: " << name << " of type ";
-
-        switch (type)
-		{
-            case DataType::INT: std::cout << "int"; break;
-            case DataType::FLOAT: std::cout << "float"; break;
-            case DataType::BOOL: std::cout << "bool"; break;
-            case DataType::STRING: std::cout << "str"; break;
-        }
-
-        std::cout << std::endl;
-        if (value)
-		{
-            printIndent(indent);
-            std::cout << "Value:" << std::endl;
-            value->print(indent + 1);
-        }
-    }
-
-    DataType getType() const
-	{
-        return type;
-    }
-
-    const char* getName() const
-	{
-        return name;
-    }
-
-    ExprNode* getValue() const
-	{
-        return value;
-    }
-
-  private:
-    DataType type;
-    const char* name;
-    ExprNode* value;
-};
-
-// If statement node
-class IfStmtNode : public StmtNode
-{
-public:
-    IfStmtNode(ExprNode* condition, StmtNode* thenBranch, StmtNode* elseBranch)
-        : condition(condition), thenBranch(thenBranch), elseBranch(elseBranch) {}
-
-    ~IfStmtNode()
-    {
-        delete condition;
-        delete thenBranch;
-        delete elseBranch;
-    }
-
-    void print(int indent = 0) const override
-    {
-        printIndent(indent);
-        std::cout << "IfStmt:" << std::endl;
-
-        printIndent(indent);
-        std::cout << "Condition:" << std::endl;
-        condition->print(indent + 1);
-
-        printIndent(indent);
-        std::cout << "Then:" << std::endl;
-        if (thenBranch) {
-            thenBranch->print(indent + 1);
-        } else {
-            printIndent(indent + 1);
-            std::cout << "<empty>" << std::endl;
-        }
-
-        if (elseBranch) {
-            printIndent(indent);
-            std::cout << "Else:" << std::endl;
-            elseBranch->print(indent + 1);
-        }
-    }
-
-    ExprNode* getCondition() const
-    {
-        return condition;
-    }
-
-    StmtNode* getThenBranch() const
-    {
-        return thenBranch;
-    }
-
-    StmtNode* getElseBranch() const
-    {
-        return elseBranch;
-    }
-
-private:
-    ExprNode* condition;
-    StmtNode* thenBranch;
-    StmtNode* elseBranch;
-};
-
-// Program node (root of the AST)
-class ProgramNode : public ASTNode
-{
-  public:
-    ~ProgramNode()
-	{
-        for (auto stmt : statements)
-		{
-            delete stmt;
-        }
-    }
-
-    void addStatement(StmtNode* stmt)
-	{
-        if (stmt)
-		{
-            statements.push_back(stmt);
-        }
-    }
-
-    void print(int indent = 0) const override
-	{
-        printIndent(indent);
-        std::cout << "Program:" << std::endl;
-        for (const auto& stmt : statements)
-		{
-            if (stmt) {
-                stmt->print(indent + 1);
-            }
-        }
-    }
-
-    const std::vector<StmtNode*>& getStatements() const
-	{
-        return statements;
-    }
-
-private:
-    std::vector<StmtNode*> statements;
-};
+// Include all AST node types
+#include "AST/ASTNode.hpp"
+#include "AST/Types.hpp"
+#include "AST/ExprNode.hpp"
+#include "AST/StmtNode.hpp"
+#include "AST/LiteralNodes.hpp"
+#include "AST/IdentifierNode.hpp"
+#include "AST/BinaryOpNode.hpp"
+#include "AST/VarDeclNode.hpp"
+#include "AST/IfStmtNode.hpp"
+#include "AST/BlockStmtNode.hpp"
+#include "AST/ProgramNode.hpp"
