+++ src/AST/StmtNode.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "ASTNode.hpp"
+
+// Statement node (base class for all statements)
+class StmtNode : public ASTNode {
+  public:
+    virtual ~StmtNode() = default;
+};
